+++ internal/dao/repo.go
@@ -160,3 +160,92 @@ from (
 where daos.id = cnt.dao_id
 `).Error
 }
+
+// GetRecommended returns the list of available dao strategies in our system
+func (r *Repo) GetRecommended() ([]Recommendation, error) {
+	query := `
+with active_daos as (select distinct dao_id
+                     from proposals pr
+                     where (
+                         to_timestamp(pr.start::double precision) >= NOW() - INTERVAL '4 month'
+                             or to_timestamp(pr.end::double precision) >= NOW() - INTERVAL '4 month'
+                         )
+                       and spam is not true)
+
+select original_id,
+       internal_id,
+       name                 strategy_name,
+       params ->> 'symbol'  symbol,
+       network              network_id,
+       params ->> 'address' address
+from (select daos.original_id,
+             daos.id        internal_id,
+             daos.network,
+             st."Name"   as name,
+             st."Params" as params
+      from daos,
+           jsonb_to_recordset(daos.strategies) AS st("Name" text, "Params" jsonb)
+      where st."Name" <> 'multichain'
+        and verified is true) data
+         inner join active_daos ad on ad.dao_id = data.internal_id
+where name in
+      ('erc20-votes', 'erc20-balance-of', 'uni',
+       'eth-balance', 'erc721', 'eth-with-balance',
+       'contract-call', 'erc1155-balance-of', 'ens-domains-owned')
+  and params ->> 'symbol' is not null
+  and params ->> 'address' is not null
+
+union all
+
+select data.original_id,
+       data.internal_id,
+       st.name as              strategy_name,
+       st.params ->> 'symbol'  symbol,
+       data.network            network_id,
+       st.params ->> 'address' address
+from (select daos.original_id,
+             daos.id        internal_id,
+             daos.network,
+             st."Name"   as name,
+             st."Params" as params
+      from (select *
+            from daos
+                     inner join active_daos ad on ad.dao_id = daos.id) daos,
+           jsonb_to_recordset(daos.strategies) AS st("Name" text, "Params" jsonb)
+      where st."Name" = 'multichain'
+        and verified is true) data,
+     jsonb_to_recordset(data.params -> 'strategies') AS st("name" text, "params" jsonb)
+where st.params ->> 'symbol' is not null
+  and st.params ->> 'address' is not null
+
+order by original_id, symbol`
+
+	rows, err := r.db.Raw(query).Rows()
+	if err != nil {
+		return nil, fmt.Errorf("get active by user: %w", err)
+	}
+
+	defer rows.Close()
+
+	defaultSize := 500
+	list := make([]Recommendation, 0, defaultSize)
+	for rows.Next() {
+		rec := Recommendation{}
+
+		err = rows.Scan(
+			&rec.OriginalId,
+			&rec.InternalId,
+			&rec.Name,
+			&rec.Symbol,
+			&rec.NetworkId,
+			&rec.Address,
+		)
+		if err != nil {
+			return nil, fmt.Errorf("convert row: %w", err)
+		}
+
+		list = append(list, rec)
+	}
+
+	return list, nil
+}